address comment: move tool emission to tool.ts
